+++ Beezle/Classes/State/Edit/EntitySelectIngameMenuState.m
@@ -44,7 +44,7 @@
 		[_menu addChild:cancelMenuItem];
 		
 		int nMenuItems = [[_menu  children] count];
-		int num1 = ceilf(nMenuItems / 3.0f);
+		int num1 = (int)ceilf(nMenuItems / 3.0f);
 		int num2 = nMenuItems - 2 * num1;
 		[_menu alignItemsInRows:[NSNumber numberWithInt:num1], [NSNumber numberWithInt:num1], [NSNumber numberWithInt:num2], nil];
